Update generate_acts.py
bf16 -> f16

+++ src/generate_acts.py
@@ -95,7 +95,7 @@ if __name__ == "__main__":
     args = parse_args()
 
     model = AutoModelForCausalLM.from_pretrained(
-        args.model_path, return_dict=True, torch_dtype=torch.bfloat16, device_map="auto"
+        args.model_path, return_dict=True, torch_dtype=torch.float16, device_map="auto"
     )
     tokenizer = AutoTokenizer.from_pretrained(args.model_path)
     model.config.use_cache = True
